Switch from buildkite to GitHub actions.
We now have the ability to connect to Packit directly from GitHub
actions, which removes the need to use our internal Buildkite
infrastructure.

The new workflow is simpler by not having to deal with Docker, and
should also be easier to debug and work with. It uses the new
`mrc-ide/orderly-action` to handle the initialization and authentication
to Packit.